Easy method to create database tables in MySQL or MariaDB
The Database Table Maker is a program written in Python that allows for easy creation of MySQL and MariaDB tables. It was originally intended for creating tables for bioinformatics applications, but it is general enough so that anybody can use it to create database tables. Data entry into tables created by Database Table Maker is easy and intuitive. Templates can be created for any table and provide default values to facilitate data entry. Actions can be called to calculate default values for fields or print labels. Modifications to any data in any table are automatically tracked by the auditor. The auditor keeps track of who modified the record, when it was modified and the value of the record before it was modified. The query module allows querying data in any table in the database. the dashboard module allows graphical display (pie, bar or line charts) of the data